requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
openai
requests
python-dotenv==1.0.0